+++ Modules/Intent.Modules.Metadata.RDBMS/Api/Extensions/AttributeModelStereotypeExtensions.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using Intent.Metadata.Models;
 using Intent.Modelers.Domain.Api;
 using Intent.Modules.Common;
-using Intent.Modules.Common.Types.Api;
 using Intent.RoslynWeaver.Attributes;
 using Intent.SdkEvolutionHelpers;